# pkg/query: Add fast map of strings to json marshaling