… pkg/query/testdata/profile-with-labels.pb.gz