@metalmatze 2023-08-07
Very cool! I knew this was eventually going to happen, but didn't expect it so soon :smiley: 
Very happy that we are not using reflection anymore now!